CLN: Move internal function to where its used
